move CustomerAutoSchema class from apps/iam to apps/engine
